fixBugs
