+++ src/components/NavBar.jsx
@@ -2,8 +2,8 @@ import { useEffect } from "react";
 import axios from "axios";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, useNavigate } from "react-router-dom";
-import { setUser} from "../store/user";
-import {setProducts } from "../store/products";
+import { setUser } from "../store/user";
+import { setProducts } from "../store/products";
 import "../Styles/NavBar.css";
 import { fakeData } from "../utils/fakeData";
 
@@ -12,28 +12,28 @@ const Navbar = () => {
   //le avisamos al store que hay un nuevo usuario logueado
   const dispatch = useDispatch();
 
-  //ahora traemos el usuario logueado con el hook useSelector
+  // //ahora traemos el usuario logueado con el hook useSelector
   const user = useSelector((state) => state.user);
   //lo que hace la funcion es traer el estado del store y lo guarda en user
 
-  useEffect(() => {
-    axios
-      .get("/api/users/me")
-      .then((res) => res.data)
-      .then((user) => {
-        console.log(user);
-        dispatch(setUser(user));
-      })
-      .catch(() =>
-        console.error("Necesitas loguearte para obtener tu usuario.")
-      );
-  }, []);
+  // useEffect(() => {}, [user]);
 
-  useEffect(() => {
-    dispatch(setProducts(fakeData));
-  }, []);
+  // useEffect(() => {
+  //   axios
+  //     .get("/api/users/me")
+  //     .then((res) => res.data)
+  //     .then((user) => {
+  //       console.log(user);
+  //       dispatch(setUser(user));
+  //     })
+  //     .catch(() =>
+  //       console.error("Necesitas loguearte para obtener tu usuario.")
+  //     );
+  // }, []);
 
-
+  // useEffect(() => {
+  //   dispatch(setProducts(fakeData));
+  // }, []);
 
   const handlerLogOut = (e) => {
     axios
